photo-server/requirements.txt
flask==3.0.3
google-cloud-storage==2.16.0
Pillow==10.3.0
pillow-heif==0.16.0
qrcode[pil]==7.4.2
gunicorn==22.0.0
